# docs/swagger.yaml
definitions:
...
  models.HttpError:
    properties:
      code:
        example: 400
        type: integer
      data:
        additionalProperties: true
        type: object
      message:
        example: status bad request
        type: string
    type: object
...
  models.Permission:
    properties:
      _id:
        type: string
      checker:
        items:
          type: number
        type: array
      endpoint:
        type: string
      maker:
        items:
          type: number
        type: array
    type: object
  permission.UpdatePermission:
    properties:
      checker:
        items:
          type: number
        type: array
      endpoint:
        type: string
      maker:
        items:
          type: number
        type: array
    type: object
info:
  contact: {}
paths:
...
  /permission/{id}:
    delete:
      description: Delete a Makerchecker Permission By ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Permission'
        "400":
          description: Id cannot be empty
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: No permission found with Id
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Delete a Makerchecker Permission by ID
      tags:
      - permission
    get:
      description: Retrieve a Makerchecker permission by Id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Permission'
        "400":
          description: Id cannot be empty
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: No permission can be found with Id
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Get Makerchecker permission by Id
      tags:
      - permission
    put:
      description: Update Makerchecker Permission by Id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/permission.UpdatePermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Permission'
        "400":
          description: Id cannot be empty and permission object is invalid
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: No permission found
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Update Makerchecker Permission by Id
      tags:
      - permission
...
swagger: "2.0"